profiler-to-carbon: set a timeout on the connection
We've noticed several times how profiler-to-carbon gets stuck and does
not report metrics to graphite anymore. This is due to this script
hanging while waiting for a response from mwprof, which evidently leaves
the connection socket open and does not closes it.

So, first and easiest part of the fix is letting profiler-to-carbon
manage those failure scenario correctly, by timing out the connection
after 60 seconds.

Change-Id: Ice3c81684b50cb7ad04e8972888d81ce4c43d0e8
Signed-off-by: Giuseppe Lavagetto <glavagetto@wikimedia.org>